% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{abbreviate_license_column}
\alias{abbreviate_license_column}
\title{Abbreviate the license column in a data frame}
\usage{
abbreviate_license_column(df, col = "license")
}
\arguments{
\item{df}{Data frame}

\item{col}{Column name to abbreviate (default: "license")}
}
\value{
Data frame with abbreviated license values when column exists
}
\description{
Abbreviate the license column in a data frame
}
