% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{add_occurrence_ids}
\alias{add_occurrence_ids}
\title{Add occurrence IDs if missing}
\usage{
add_occurrence_ids(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with occurrenceID column
}
\description{
Add occurrence IDs if missing
}
