% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{clean_coordinate_separators}
\alias{clean_coordinate_separators}
\title{Clean coordinate separators (comma to dot)}
\usage{
clean_coordinate_separators(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with cleaned coordinates
}
\description{
Clean coordinate separators (comma to dot)
}
