% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{clean_scientific_names}
\alias{clean_scientific_names}
\title{Validate and clean scientific names}
\usage{
clean_scientific_names(names_vector)
}
\arguments{
\item{names_vector}{Character vector of scientific names}
}
\value{
Character vector of cleaned names
}
\description{
Validate and clean scientific names
}
