% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{detect_delimiter}
\alias{detect_delimiter}
\title{Detect CSV delimiter}
\usage{
detect_delimiter(file_path)
}
\arguments{
\item{file_path}{Path to CSV file}
}
\value{
Delimiter character
}
\description{
Detect CSV delimiter
}
