% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{detect_encoding}
\alias{detect_encoding}
\title{Detect file encoding}
\usage{
detect_encoding(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
Detected encoding string
}
\description{
Detect file encoding
}
