% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{fix_dates_to_iso}
\alias{fix_dates_to_iso}
\title{Convert dates to ISO 8601 format}
\usage{
fix_dates_to_iso(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with converted dates
}
\description{
Convert dates to ISO 8601 format
}
