% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_i18n.R
\name{get_language_name}
\alias{get_language_name}
\title{Get language display name}
\usage{
get_language_name(lang_code)
}
\arguments{
\item{lang_code}{Character. Language code ("pt" or "en")}
}
\value{
Character. Display name
}
\description{
Get language display name
}
