% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dwc.R
\name{get_required_dwc_terms}
\alias{get_required_dwc_terms}
\title{Get required Darwin Core terms}
\usage{
get_required_dwc_terms()
}
\value{
Data frame with required DwC terms
}
\description{
Get required Darwin Core terms
}
