% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{i18n_dict}
\alias{i18n_dict}
\title{Translation Dictionary}
\format{
List of named lists with pt and en keys
}
\usage{
i18n_dict
}
\description{
Named list containing all UI strings in Portuguese (pt) and English (en)
}
\keyword{datasets}
