% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dwc.R
\name{load_dwc_terms_rds}
\alias{load_dwc_terms_rds}
\title{Load Darwin Core terms from RDS}
\usage{
load_dwc_terms_rds(force = FALSE)
}
\arguments{
\item{force}{Logical scalar. If TRUE, invalidate cache and reload from disk.}
}
\value{
Data frame with DwC terms
}
\description{
Load Darwin Core terms from RDS
}
