% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help.R
\name{mod_help_server}
\alias{mod_help_server}
\title{Help Module Server}
\usage{
mod_help_server(id, lang_r)
}
\arguments{
\item{id}{Module ID}

\item{lang_r}{Reactive language value}
}
\description{
Help Module Server
}
