% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preview.R
\name{mod_preview_server}
\alias{mod_preview_server}
\title{Preview Module Server}
\usage{
mod_preview_server(id, mapped_data_r, lang_r)
}
\arguments{
\item{id}{Module ID}

\item{mapped_data_r}{Reactive data frame with mapped data}

\item{lang_r}{Reactive language value}
}
\value{
Reactive preview data frame
}
\description{
Preview Module Server
}
