% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Upload Module Server}
\usage{
mod_upload_server(id, lang_r)
}
\arguments{
\item{id}{Module ID}

\item{lang_r}{Reactive language value}
}
\value{
Reactive data frame with raw data
}
\description{
Upload Module Server
}
