% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validate_coords.R
\name{mod_validate_coords_ui}
\alias{mod_validate_coords_ui}
\title{Validate Coordinates Module UI}
\usage{
mod_validate_coords_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Validate Coordinates Module UI
}
