% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validate_names.R
\name{mod_validate_names_ui}
\alias{mod_validate_names_ui}
\title{Validate Names Module UI}
\usage{
mod_validate_names_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Validate Names Module UI
}
