% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wiki.R
\name{mod_wiki_server}
\alias{mod_wiki_server}
\title{Wiki Module Server}
\usage{
mod_wiki_server(id, lang_r)
}
\arguments{
\item{id}{Module ID}

\item{lang_r}{Reactive language value}
}
\description{
Wiki Module Server
}
