% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{process_for_export}
\alias{process_for_export}
\title{Process data for DwC-compliant export}
\usage{
process_for_export(df)
}
\arguments{
\item{df}{Data frame with mapped columns}
}
\value{
Data frame with ISO dates, cleaned separators, UUIDs
}
\description{
Process data for DwC-compliant export
}
