% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{strip_bom}
\alias{strip_bom}
\title{Strip UTF-8 BOM from text}
\usage{
strip_bom(text)
}
\arguments{
\item{text}{Character vector}
}
\value{
Character vector without BOM prefix
}
\description{
Strip UTF-8 BOM from text
}
