% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_i18n.R
\name{tr}
\alias{tr}
\title{Translate a key to current language}
\usage{
tr(key, lang = "en")
}
\arguments{
\item{key}{Character. Key from i18n_dict}

\item{lang}{Character. "pt" or "en"}
}
\value{
Character. Translated string
}
\description{
Translate a key to current language
}
