% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_upload.R
\name{upload_camtrap_requirements_ui}
\alias{upload_camtrap_requirements_ui}
\title{Build the Camtrap mode "format requirements" body}
\usage{
upload_camtrap_requirements_ui(lang)
}
\arguments{
\item{lang}{Language code, "pt" or "en".}
}
\value{
A `shiny.tag` representing the file-row list for Camtrap DP and
  Wildlife Insights expected files.
}
\description{
Build the Camtrap mode "format requirements" body
}
\keyword{internal}
