% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dwc.R
\name{validate_coords}
\alias{validate_coords}
\title{Validate coordinates (WGS84)}
\usage{
validate_coords(lat, lon)
}
\arguments{
\item{lat}{Numeric vector of latitudes}

\item{lon}{Numeric vector of longitudes}
}
\value{
Data frame with validation results
}
\description{
Validate coordinates (WGS84)
}
