% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dwc.R
\name{validate_occurrence_id}
\alias{validate_occurrence_id}
\title{Check for unique occurrence IDs}
\usage{
validate_occurrence_id(ids)
}
\arguments{
\item{ids}{Character vector of occurrence IDs}
}
\value{
Logical vector indicating uniqueness
}
\description{
Check for unique occurrence IDs
}
