% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export.R
\name{abbreviate_license}
\alias{abbreviate_license}
\title{Abbreviate Creative Commons license values}
\usage{
abbreviate_license(x)
}
\arguments{
\item{x}{Character vector with license values}
}
\value{
Character vector with known license URLs abbreviated
}
\description{
Abbreviate Creative Commons license values
}
