% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapping.R
\name{mod_mapping_server}
\alias{mod_mapping_server}
\title{Mapping Module Server}
\usage{
mod_mapping_server(id, raw_data_r, lang_r)
}
\arguments{
\item{id}{Module namespace ID}

\item{raw_data_r}{Reactive data frame from upload module}

\item{lang_r}{Reactive language code}
}
\value{
Reactive containing processed/mapped data
}
\description{
Mapping Module Server
}
