% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preview.R
\name{mod_preview_server}
\alias{mod_preview_server}
\title{Preview Module Server}
\usage{
mod_preview_server(
  id,
  mapped_data_r,
  lang_r,
  download_data_r = mapped_data_r,
  name_review_payload_r = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{mapped_data_r}{Reactive data frame with mapped data}

\item{lang_r}{Reactive language value}

\item{download_data_r}{Reactive data frame with full mapped data for download}

\item{name_review_payload_r}{Optional reactive payload from name manual review}
}
\value{
Reactive preview data frame
}
\description{
Preview Module Server
}
