% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validate_names.R
\name{mod_validate_names_server}
\alias{mod_validate_names_server}
\title{Validate Names Module Server}
\usage{
mod_validate_names_server(id, mapped_data_r, lang_r)
}
\arguments{
\item{id}{Module ID}

\item{mapped_data_r}{Reactive data frame with mapped data}

\item{lang_r}{Reactive language value}
}
\description{
Validate Names Module Server
}
