% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Saira Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to shinyApp}
}
\value{
A Shiny app object
}
\description{
Starts the Saira app for biodiversity data standardization
}
