% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saira-package.R
\docType{package}
\name{saira-package}
\alias{saira}
\alias{saira-package}
\title{saira: Biodiversity Data Standardization to Darwin Core}
\description{
Shiny application for standardizing biodiversity datasets to the Darwin Core standard. Provides bilingual interface (PT-BR/EN-US) with tools for data validation, column mapping, and taxonomic verification. Designed for use with SiBBr (Sistema de Informação sobre a Biodiversidade Brasileira).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rogerio-onza/saira}
  \item Report bugs at \url{https://github.com/rogerio-onza/saira/issues}
}

}
\author{
\strong{Maintainer}: Rogério Nunes Oliveira \email{rogerio@sibbr.gov.br}

}
\keyword{internal}
